+++ sentry/src/main/java/io/sentry/Integration.java
@@ -1,5 +1,7 @@
 package io.sentry;
 
+import org.jetbrains.annotations.NotNull;
+
 /**
  * Code that provides middlewares, bindings or hooks into certain frameworks or environments, along
  * with code that inserts those bindings and activates them.
@@ -11,5 +13,5 @@ public interface Integration {
    * @param hub the Hub
    * @param options the options
    */
-  void register(IHub hub, SentryOptions options);
+  void register(@NotNull IHub hub, @NotNull SentryOptions options);
 }