+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Bump: sentry-native to 0.4.7
 * Enchancement: Add ability to flush events synchronously.
 * Fix: Disable Gson HTML escaping
+* Enchancement: Add overloads for startTrasaction taking op and description (#1244)
 * Enchancement: Support @SentrySpan and @SentryTransaction on classes and interfaces. (#1243)
 * Enchancement: Do not serialize empty collections and maps (#1245)
 * Ref: Simplify RestTemplate instrumentation (#1246)