+++ CHANGELOG.md
@@ -6,7 +6,10 @@
 * Enchancement: Support @SentrySpan and @SentryTransaction on classes and interfaces. (#1243)
 * Enchancement: Do not serialize empty collections and maps (#1245)
 * Ref: Simplify RestTemplate instrumentation (#1246)
+
+Breaking Changes:
 * Enchancement: SentryExceptionResolver should not send handled errors by default (#1248).
+
 
 # 4.1.0