+++ CHANGELOG.md
@@ -8,10 +8,10 @@
 * Enchancement: Support @SentrySpan and @SentryTransaction on classes and interfaces. (#1243)
 * Enchancement: Do not serialize empty collections and maps (#1245)
 * Ref: Simplify RestTemplate instrumentation (#1246)
-* Enchancement: Add overloads for startTrasaction taking op and description (#1244)
 
 Breaking Changes:
 * Enchancement: SentryExceptionResolver should not send handled errors by default (#1248).
+* Enchancement: Add overloads for startTransaction taking op and description (#1244)
 
 
 # 4.1.0